Refactor: parsing code extracted into separate modules.
