+++ src/formats/pro/parse.rs
@@ -1,3 +1,5 @@
+mod r#type;
+
 use super::*;
 
 use std::convert::TryInto;
@@ -37,12 +39,10 @@ pub fn prototype<S: Read + Seek>(source: &mut S) -> Result<Prototype, errors::Er
         Ok(value) => value,
     };
 
-    let r#type = id_bytes[0];
-
-    let object_id_bytes = &id_bytes[2..4];
-    let object_id = u16::from_be_bytes(match object_id_bytes.try_into() {
+    let type_id = id_bytes[0];
+    let object_id = u16::from_be_bytes(match &id_bytes[2..4].try_into() {
         Err(_) => return Err(errors::Error::Source),
-        Ok(value) => value,
+        Ok(value) => *value,
     });
 
     let mut text_id_bytes = [0u8; 4];
@@ -65,7 +65,7 @@ pub fn prototype<S: Read + Seek>(source: &mut S) -> Result<Prototype, errors::Er
         Ok(value) => value,
     };
 
-    let lradius = u32::from_be_bytes(lradius_bytes);
+    let lradius = lradius_bytes[3];
 
     let mut lintensity_bytes = [0u8; 4];
     match source.read_exact(&mut lintensity_bytes) {
@@ -73,7 +73,10 @@ pub fn prototype<S: Read + Seek>(source: &mut S) -> Result<Prototype, errors::Er
         Ok(value) => value,
     };
 
-    let lintensity = u32::from_be_bytes(lintensity_bytes);
+    let lintensity = u16::from_be_bytes(match &lintensity_bytes[2..4].try_into() {
+        Ok(value) => *value,
+        Err(_) => return Err(errors::Error::Source),
+    });
 
     let mut flags_bytes = [0u8; 4];
     match source.read_exact(&mut flags_bytes) {
@@ -157,1068 +160,30 @@ pub fn prototype<S: Read + Seek>(source: &mut S) -> Result<Prototype, errors::Er
         return Err(errors::Error::Format(errors::Format::Flags));
     }
 
-    let meta = meta::Info {
-        light: meta::info::Light {
-            distance: ScaledValue {
-                value: match u8::try_from(lradius) {
-                    Ok(value) => value,
-                    Err(_) => return Err(errors::Error::Format(errors::Format::Data))
-                },
-                scale: Range { start: 0, end: 8 },
-            },
-            intensity: ScaledValue {
-                value: match u16::try_from(lintensity) {
-                    Ok(value) => value,
-                    Err(_) => return Err(errors::Error::Format(errors::Format::Data))
-                },
-                scale: Range { start: 0, end: u16::MAX },
-            },
-        },
-        flags: flagset,
-        sprite: match object::common::sprite::Reference::try_from(sprite_id_bytes) {
-            Ok(value) => value,
-            Err(_) => return Err(errors::Error::Format(errors::Format::Data))
-        },
-        connections: meta::info::Connections {
-            description_id: text_id,
-        },
-    };
-
-    let object_type = match r#type {
-        0 => {
-            let mut item_flags_bytes = [0u8; 3];
-            match source.read_exact(&mut item_flags_bytes[1..=3]) {
-                Err(error) => return Err(errors::Error::Read(error)),
-                Ok(value) => value,
-            };
-
-            let mut item_flags: HashSet<object::item::Flag> = HashSet::new();
-            let mut weapon_flags: HashSet<object::item::weapon::Flag> = HashSet::new();
-            let mut item_actions: HashSet<object::common::actions::Instance> = HashSet::new();
-
-            if (item_flags_bytes[0] & 0x01) == 0x01 {
-                if !weapon_flags.insert(object::item::weapon::Flag::BigGun) {
-                    return Err(errors::Error::Format(errors::Format::Flags));
-                }
-            }
-
-            if (item_flags_bytes[0] & 0x02) == 0x02 {
-                if !weapon_flags.insert(object::item::weapon::Flag::SecondHand) {
-                    return Err(errors::Error::Format(errors::Format::Flags));
-                }
-            }
-
-            if (item_flags_bytes[0] & 0x80) == 0x80 {
-                if !item_actions.insert(object::common::actions::Instance::PickUp) {
-                    return Err(errors::Error::Format(errors::Format::Flags));
-                }
-            }
-
-            if (item_flags_bytes[2] & 0x08) == 0x08 {
-                if !item_flags.insert(object::item::Flag::Hidden) {
-                    return Err(errors::Error::Format(errors::Format::Flags));
-                }
-            }
-
-            let can_use = (item_flags_bytes[0] & 0x08) == 0x08;
-            let can_use_on = (item_flags_bytes[0] & 0x10) == 0x10;
-
-            let usage = object::common::actions::Usage {
-                itself: can_use,
-                something: can_use_on,
-                knees_down: false,
-            };
-
-            if !item_actions.insert(object::common::actions::Instance::Usage(usage)) {
-                return Err(errors::Error::Format(errors::Format::Flags));
-            }
-
-            let mut attack_modes_bytes = [0u8; 1];
-            match source.read_exact(&mut attack_modes_bytes) {
-                Err(error) => return Err(errors::Error::Read(error)),
-                Ok(value) => value,
-            };
-
-            let attack_modes = u8::from_be_bytes(attack_modes_bytes);
-
-            let attack_mode_primary_raw = attack_modes & 0xf;
-            let attack_mode_secondary_raw = (attack_modes >> 4) & 0xf;
-
-            let attack_mode_primary =
-                match attack_mode_primary_raw {
-                    0 => None,
-                    value => Some(
-                        match object::item::weapon::attack::Mode::try_from(value) {
-                            Ok(value) => value,
-                            Err(_) => return Err(errors::Error::Format(errors::Format::Data))
-                        }
-                    )
-                };
-
-            let attack_mode_secondary =
-                match attack_mode_secondary_raw {
-                    0 => None,
-                    value => Some(
-                        match object::item::weapon::attack::Mode::try_from(value) {
-                            Ok(value) => value,
-                            Err(_) => return Err(errors::Error::Format(errors::Format::Data))
-                        }
-                    )
-                };
-
-            let mut item_script_id_bytes = [0u8; 4];
-            match source.read_exact(&mut item_script_id_bytes) {
-                Err(error) => return Err(errors::Error::Read(error)),
-                Ok(value) => value,
-            };
-
-            let mut item_type_bytes = [0u8; 4];
-            match source.read_exact(&mut item_type_bytes) {
-                Err(error) => return Err(errors::Error::Read(error)),
-                Ok(value) => value,
-            };
-
-            let item_type = u32::from_be_bytes(item_type_bytes);
-
-            let mut material_id_bytes = [0u8; 4];
-            match source.read_exact(&mut material_id_bytes) {
-                Err(error) => return Err(errors::Error::Read(error)),
-                Ok(value) => value,
-            };
-
-            let material_id = u32::from_be_bytes(material_id_bytes);
-
-            let mut item_size_bytes = [0u8; 4];
-            match source.read_exact(&mut item_size_bytes) {
-                Err(error) => return Err(errors::Error::Read(error)),
-                Ok(value) => value,
-            };
-
-            let item_size = u32::from_be_bytes(item_size_bytes);
-
-            let mut item_weight_bytes = [0u8; 4];
-            match source.read_exact(&mut item_weight_bytes) {
-                Err(error) => return Err(errors::Error::Read(error)),
-                Ok(value) => value,
-            };
-
-            let item_weight = u32::from_be_bytes(item_weight_bytes);
-
-            let mut item_cost_bytes = [0u8; 4];
-            match source.read_exact(&mut item_cost_bytes) {
-                Err(error) => return Err(errors::Error::Read(error)),
-                Ok(value) => value,
-            };
-
-            let item_cost = u32::from_be_bytes(item_cost_bytes);
-
-            let mut item_sprite_id_bytes = [0u8; 4];
-            match source.read_exact(&mut item_sprite_id_bytes) {
-                Err(error) => return Err(errors::Error::Read(error)),
-                Ok(value) => value,
-            };
-
-            let mut item_sound_ids_bytes = [0u8; 1];
-            match source.read_exact(&mut item_sound_ids_bytes) {
-                Err(error) => return Err(errors::Error::Read(error)),
-                Ok(value) => value,
-            };
-
-            let item_sound_ids = u8::from_be_bytes(item_sound_ids_bytes);
-
-            let item_type = match item_type {
-                0 => {
-                    let mut armor_ac_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_ac_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_ac = u32::from_be_bytes(armor_ac_bytes);
-
-                    let mut armor_dr_normal_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dr_normal_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dr_normal = u32::from_be_bytes(armor_dr_normal_bytes);
-
-                    let mut armor_dr_laser_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dr_laser_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dr_laser = u32::from_be_bytes(armor_dr_laser_bytes);
-
-                    let mut armor_dr_fire_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dr_fire_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dr_fire = u32::from_be_bytes(armor_dr_fire_bytes);
-
-                    let mut armor_dr_plasma_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dr_plasma_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dr_plasma = u32::from_be_bytes(armor_dr_plasma_bytes);
-
-                    let mut armor_dr_electrical_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dr_electrical_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dr_electrical = u32::from_be_bytes(armor_dr_electrical_bytes);
-
-                    let mut armor_dr_emp_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dr_emp_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dr_emp = u32::from_be_bytes(armor_dr_emp_bytes);
-
-                    let mut armor_dr_explosive_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dr_explosive_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dr_explosive = u32::from_be_bytes(armor_dr_explosive_bytes);
-
-                    let mut armor_dt_normal_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dt_normal_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dt_normal = u32::from_be_bytes(armor_dt_normal_bytes);
-
-                    let mut armor_dt_laser_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dt_laser_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dt_laser = u32::from_be_bytes(armor_dt_laser_bytes);
-
-                    let mut armor_dt_fire_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dt_fire_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dt_fire = u32::from_be_bytes(armor_dt_fire_bytes);
-
-                    let mut armor_dt_plasma_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dt_plasma_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dt_plasma = u32::from_be_bytes(armor_dt_plasma_bytes);
-
-                    let mut armor_dt_electrical_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dt_electrical_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dt_electrical = u32::from_be_bytes(armor_dt_electrical_bytes);
-
-                    let mut armor_dt_emp_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dt_emp_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dt_emp = u32::from_be_bytes(armor_dt_emp_bytes);
-
-                    let mut armor_dt_explosive_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_dt_explosive_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_dt_explosive = u32::from_be_bytes(armor_dt_explosive_bytes);
-
-                    let mut armor_perk_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_perk_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let armor_perk = i32::from_be_bytes(armor_perk_bytes);
-
-                    let mut armor_male_fid_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_male_fid_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let mut armor_female_fid_bytes = [0u8; 4];
-                    match source.read_exact(&mut armor_female_fid_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    object::item::Type::Armor(
-                        object::item::armor::Instance {
-                            class: armor_ac,
-                            threshold: HashMap::from([
-                                (object::common::combat::damage::Type::Default, armor_dt_normal),
-                                (object::common::combat::damage::Type::Laser, armor_dt_laser),
-                                (object::common::combat::damage::Type::Fire, armor_dt_fire),
-                                (object::common::combat::damage::Type::Plasma, armor_dt_plasma),
-                                (object::common::combat::damage::Type::Electrical, armor_dt_electrical),
-                                (object::common::combat::damage::Type::Emp, armor_dt_emp),
-                                (object::common::combat::damage::Type::Explosive, armor_dt_explosive),
-                            ]),
-                            resistance: HashMap::from([
-                                (object::common::combat::damage::Type::Default, armor_dr_normal),
-                                (object::common::combat::damage::Type::Laser, armor_dr_laser),
-                                (object::common::combat::damage::Type::Fire, armor_dr_fire),
-                                (object::common::combat::damage::Type::Plasma, armor_dr_plasma),
-                                (object::common::combat::damage::Type::Electrical, armor_dr_electrical),
-                                (object::common::combat::damage::Type::Emp, armor_dr_emp),
-                                (object::common::combat::damage::Type::Explosive, armor_dr_explosive),
-                            ]),
-                            perk: match armor_perk {
-                                -1 => Option::None,
-                                value => Option::Some(
-                                    match object::common::critter::Perk::try_from(value) {
-                                        Ok(value) => value,
-                                        Err(_) =>
-                                            return Err(errors::Error::Format(errors::Format::Data))
-                                    }
-                                ),
-                            },
-                            appearance: object::item::armor::Appearance {
-                                sprites: HashMap::from([
-                                    (object::common::critter::Gender::Male,
-                                     match object::common::sprite::Reference::try_from(armor_male_fid_bytes) {
-                                         Ok(value) => value,
-                                         Err(_) => return Err(errors::Error::Format(errors::Format::Data)),
-                                     }),
-                                    (object::common::critter::Gender::Female,
-                                     match object::common::sprite::Reference::try_from(armor_female_fid_bytes) {
-                                         Ok(value) => value,
-                                         Err(_) => return Err(errors::Error::Format(errors::Format::Data)),
-                                     })
-                                ])
-                            },
-                        }
-                    )
-                }
-                1 => {
-                    let mut container_size_bytes = [0u8; 4];
-                    match source.read_exact(&mut container_size_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let container_size = u32::from_be_bytes(container_size_bytes);
-
-                    let mut container_flags_bytes = [0u8; 4];
-                    match source.read_exact(&mut container_flags_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let mut container_flags: HashSet<object::item::container::Flag> =
-                        HashSet::new();
-
-                    if (container_flags_bytes[3] & 0x01) == 0x01 {
-                        if !container_flags.insert(object::item::container::Flag::NoPickUp) {
-                            return Err(errors::Error::Format(errors::Format::Flags));
-                        }
-                    }
-
-                    if (container_flags_bytes[3] & 0x08) == 0x08 {
-                        if !container_flags.insert(object::item::container::Flag::MagicHands) {
-                            return Err(errors::Error::Format(errors::Format::Flags));
-                        }
-                    }
-
-                    object::item::Type::Container(
-                        object::item::container::Instance {
-                            size: container_size,
-                            flags: container_flags,
-                        }
-                    )
-                }
-                2 => {
-                    let mut drug_stat0_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_stat0_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_stat0_raw = i32::from_be_bytes(drug_stat0_bytes);
-
-                    let mut drug_stat1_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_stat1_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_stat1_raw = i32::from_be_bytes(drug_stat1_bytes);
-
-                    let mut drug_stat2_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_stat2_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_stat2_raw = i32::from_be_bytes(drug_stat2_bytes);
-
-                    let mut drug_effect0_amount0_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_effect0_amount0_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_effect0_amount0_raw = u32::from_be_bytes(drug_effect0_amount0_bytes);
-
-                    let mut drug_effect0_amount1_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_effect0_amount1_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_effect0_amount1_raw = u32::from_be_bytes(drug_effect0_amount1_bytes);
-
-                    let mut drug_effect0_amount2_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_effect0_amount2_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_effect0_amount2_raw = u32::from_be_bytes(drug_effect0_amount2_bytes);
-
-                    let mut drug_effect1_duration_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_effect1_duration_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_effect1_duration_raw = u32::from_be_bytes(drug_effect1_duration_bytes);
-
-                    let mut drug_effect1_amount0_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_effect1_amount0_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_effect1_amount0_raw = u32::from_be_bytes(drug_effect1_amount0_bytes);
-
-                    let mut drug_effect1_amount1_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_effect1_amount1_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_effect1_amount1_raw = u32::from_be_bytes(drug_effect1_amount1_bytes);
-
-                    let mut drug_effect1_amount2_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_effect1_amount2_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_effect1_amount2_raw = u32::from_be_bytes(drug_effect1_amount2_bytes);
-
-                    let mut drug_effect2_duration_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_effect2_duration_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_effect2_duration_raw = u32::from_be_bytes(drug_effect2_duration_bytes);
-
-                    let mut drug_effect2_amount0_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_effect2_amount0_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_effect2_amount0_raw = u32::from_be_bytes(drug_effect2_amount0_bytes);
-
-                    let mut drug_effect2_amount1_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_effect2_amount1_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_effect2_amount1_raw = u32::from_be_bytes(drug_effect2_amount1_bytes);
-
-                    let mut drug_effect2_amount2_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_effect2_amount2_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_effect2_amount2_raw = u32::from_be_bytes(drug_effect2_amount2_bytes);
-
-                    let mut drug_addiction_rate_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_addiction_rate_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_addiction_rate_raw = u32::from_be_bytes(drug_addiction_rate_bytes);
-
-                    let mut drug_addiction_perk_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_addiction_perk_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_addiction_perk_raw = i32::from_be_bytes(drug_addiction_perk_bytes);
-
-                    let mut drug_addiction_delay_bytes = [0u8; 4];
-                    match source.read_exact(&mut drug_addiction_delay_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let drug_addiction_delay_raw = u32::from_be_bytes(drug_addiction_delay_bytes);
-
-                    let statistic0 =
-                        match drug_stat0_raw > 0 {
-                            false => Option::None,
-                            true => Option::Some(
-                                match
-                                object::common::critter::Statistic::try_from(drug_stat0_raw) {
-                                    Ok(value) => value,
-                                    Err(_) =>
-                                        return Err(errors::Error::Format(errors::Format::Data))
-                                }
-                            )
-                        };
-
-                    let statistic1 =
-                        match object::common::critter::Statistic::try_from(drug_stat1_raw) {
-                            Ok(value) => value,
-                            Err(_) => return Err(errors::Error::Format(errors::Format::Data))
-                        };
-
-                    let statistic2 =
-                        match object::common::critter::Statistic::try_from(drug_stat2_raw) {
-                            Ok(value) => value,
-                            Err(_) => return Err(errors::Error::Format(errors::Format::Data))
-                        };
-
-                    let mut impact0 = HashMap::from([
-                        (statistic1.clone(), match statistic0 {
-                            Some(_) => object::item::drug::Amount::Fixed(drug_effect0_amount1_raw),
-                            None => object::item::drug::Amount::Random(
-                                drug_effect0_amount0_raw..=drug_effect0_amount1_raw
-                            ),
-                        }),
-                        (statistic2.clone(), object::item::drug::Amount::Fixed(drug_effect0_amount2_raw))
-                    ]);
-
-                    if let Some(value) = statistic0.clone() {
-                        impact0.entry(value.clone()).or_insert(
-                            object::item::drug::Amount::Fixed(
-                                drug_effect0_amount0_raw
-                            ));
-                    }
-
-                    let effect0 = object::item::drug::Effect { delay: None, impact: impact0 };
-
-                    let mut impact1 = HashMap::from([
-                        (statistic1.clone(), match statistic0 {
-                            Some(_) => object::item::drug::Amount::Fixed(drug_effect1_amount1_raw),
-                            None => object::item::drug::Amount::Random(
-                                drug_effect1_amount0_raw..=drug_effect1_amount1_raw
-                            ),
-                        }),
-                        (statistic2.clone(), object::item::drug::Amount::Fixed(drug_effect1_amount2_raw))
-                    ]);
-
-                    if let Some(value) = statistic0.clone() {
-                        impact1.entry(value.clone()).or_insert(
-                            object::item::drug::Amount::Fixed(drug_effect1_amount0_raw)
-                        );
-                    }
-
-                    let effect1 = object::item::drug::Effect {
-                        delay: Some(Duration::new(drug_effect1_duration_raw as u64 * 60, 0)),
-                        impact: impact1,
-                    };
-
-                    let mut impact2 = HashMap::from([
-                        (statistic1.clone(), match statistic0 {
-                            Some(_) => object::item::drug::Amount::Fixed(drug_effect2_amount1_raw),
-                            None => object::item::drug::Amount::Random(
-                                drug_effect2_amount0_raw..=drug_effect2_amount1_raw
-                            ),
-                        }),
-                        (statistic2.clone(), object::item::drug::Amount::Fixed(drug_effect2_amount2_raw))
-                    ]);
-
-                    if let Some(value) = statistic0.clone() {
-                        impact2.entry(value.clone()).or_insert(
-                            object::item::drug::Amount::Fixed(drug_effect2_amount0_raw)
-                        );
-                    }
-
-                    let effect2 = object::item::drug::Effect {
-                        delay: Some(Duration::new(drug_effect2_duration_raw as u64 * 60, 0)),
-                        impact: impact2,
-                    };
-
-                    if drug_effect2_duration_raw <= drug_effect1_duration_raw {
-                        return Err(errors::Error::Format(errors::Format::Consistency));
-                    }
-
-                    let addiction_perk = match drug_addiction_perk_raw {
-                        -1 => None,
-                        value => Some(
-                            match object::common::critter::Perk::
-                            try_from(value) {
-                                Ok(value) => value,
-                                Err(_) => return Err(errors::Error::Format(errors::Format::Data)),
-                            }
-                        )
-                    };
-
-                    object::item::Type::Drug(
-                        object::item::drug::Instance {
-                            effects: [effect0, effect1, effect2],
-                            addiction: match addiction_perk {
-                                None => None,
-                                Some(perk) => Some(object::item::drug::Addiction {
-                                    perk,
-                                    delay: Duration::new(
-                                        drug_addiction_delay_raw as u64 * 60, 0,
-                                    ),
-                                    chance: ScaledValue {
-                                        value: drug_addiction_rate_raw as u8,
-                                        scale: 0u8..101u8,
-                                    },
-                                })
-                            },
-                        }
-                    )
-                }
-                3 => {
-                    let mut weapon_animation_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_animation_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_animation_raw = u32::from_be_bytes(weapon_animation_bytes);
-
-                    let weapon_animation = match weapon_animation_raw {
-                        0x00 => None,
-                        value => Some(
-                            match object::item::weapon::Animation::try_from(value) {
-                                Err(_) => return Err(errors::Error::Source),
-                                Ok(value) => value,
-                            }
-                        )
-                    };
-
-                    let mut weapon_min_dmg_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_min_dmg_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_min_dmg = u32::from_be_bytes(weapon_min_dmg_bytes);
-
-                    let mut weapon_max_dmg_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_max_dmg_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_max_dmg = u32::from_be_bytes(weapon_max_dmg_bytes);
-
-                    let mut weapon_dmg_type_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_dmg_type_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_dmg_type_raw = u32::from_be_bytes(weapon_dmg_type_bytes);
-
-                    let weapon_dmg_type = match object::common::combat::damage::Type::try_from(
-                        weapon_dmg_type_raw as u8
-                    ) {
-                        Ok(value) => value,
-                        Err(_) => return Err(errors::Error::Format(errors::Format::Data)),
-                    };
-
-                    let weapon_damage = object::item::weapon::Damage {
-                        value: weapon_min_dmg..=weapon_max_dmg,
-                        r#type: weapon_dmg_type,
-                    };
-
-                    let mut weapon_dmg_range_max1_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_dmg_range_max1_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_dmg_range_max1 = u32::from_be_bytes(weapon_dmg_range_max1_bytes);
-
-                    let mut weapon_dmg_range_max2_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_dmg_range_max2_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_dmg_range_max2 = u32::from_be_bytes(weapon_dmg_range_max2_bytes);
-
-                    let mut weapon_projectile_header_bytes = [0u8; 2];
-                    match source.read_exact(&mut weapon_projectile_header_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_projectile_header = u16::from_be_bytes(weapon_projectile_header_bytes);
-
-                    if 0x0500 != weapon_projectile_header {
-                        return Err(errors::Error::Format(errors::Format::Consistency));
-                    }
-
-                    let mut weapon_projectile_idx_bytes = [0u8; 2];
-                    match source.read_exact(&mut weapon_projectile_idx_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_projectile_idx = u16::from_be_bytes(weapon_projectile_idx_bytes);
-
-                    let mut weapon_min_strength_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_min_strength_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_min_strength = u32::from_be_bytes(weapon_min_strength_bytes);
-
-                    let mut weapon_cost1_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_cost1_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_cost1 = u32::from_be_bytes(weapon_cost1_bytes);
-
-                    let mut weapon_cost2_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_cost2_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_cost2 = u32::from_be_bytes(weapon_cost2_bytes);
-
-                    let weapon_attack1 = object::item::weapon::attack::Instance {
-                        cost: weapon_cost1,
-                        mode: attack_mode_primary,
-                        range: 0..=weapon_dmg_range_max1,
-                    };
-
-                    let weapon_attack2 = object::item::weapon::attack::Instance {
-                        cost: weapon_cost2,
-                        mode: attack_mode_secondary,
-                        range: 0..=weapon_dmg_range_max2,
-                    };
-
-                    let mut weapon_crit_list_idx_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_crit_list_idx_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_crit_list_idx = u32::from_be_bytes(weapon_crit_list_idx_bytes);
-
-                    let mut weapon_perk_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_perk_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_perk_raw = i32::from_be_bytes(weapon_perk_bytes);
-
-                    let weapon_perk = match weapon_perk_raw {
-                        -1 => Option::None,
-                        value => Option::Some(
-                            match object::common::critter::Perk::try_from(value) {
-                                Ok(value) => value,
-                                Err(_) =>
-                                    return Err(errors::Error::Format(errors::Format::Data))
-                            }
-                        ),
-                    };
-
-                    let mut weapon_burst_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_burst_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_burst_count = u32::from_be_bytes(weapon_burst_bytes);
-
-                    let mut weapon_caliber_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_caliber_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_caliber_raw = u32::from_be_bytes(weapon_caliber_bytes);
-
-                    let weapon_caliber =
-                        match weapon_caliber_raw {
-                            0 => None,
-                            value => Some(
-                                match object::common::weapons::Caliber::try_from(value) {
-                                    Ok(value) => value,
-                                    Err(_) => return Err(errors::Error::Format(errors::Format::Data))
-                                }
-                            )
-                        };
-
-                    let mut weapon_ammo_pid_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_ammo_pid_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_ammo_pid = u32::from_be_bytes(weapon_ammo_pid_bytes);
-
-                    let mut weapon_capacity_bytes = [0u8; 4];
-                    match source.read_exact(&mut weapon_capacity_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_capacity = u32::from_be_bytes(weapon_capacity_bytes);
-
-                    let mut weapon_sound_ids_bytes = [0u8; 1];
-                    match source.read_exact(&mut weapon_sound_ids_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let weapon_sound_ids = u8::from_be_bytes(weapon_sound_ids_bytes);
-
-                    object::item::Type::Weapon(object::item::weapon::Instance {
-                        flags: weapon_flags,
-                        damage: weapon_damage,
-                        attacks: [weapon_attack1, weapon_attack2],
-                        animation: weapon_animation,
-                        requirements: object::item::weapon::Requirements {
-                            strength: weapon_min_strength
-                        },
-                        rounds: object::item::weapon::Rounds {
-                            burst: weapon_burst_count,
-                            magazine: weapon_capacity,
-                        },
-                        caliber: weapon_caliber,
-                        perk: weapon_perk,
-                        connections: object::item::weapon::Connections {
-                            ammo_item_id: weapon_ammo_pid,
-                            failure_list_id: weapon_crit_list_idx,
-                            projectile_misc_id: weapon_projectile_idx,
-                            _sounds_ids: weapon_sound_ids,
-                        },
-                    })
-                }
-                4 => {
-                    let mut ammo_caliber_bytes = [0u8; 4];
-                    match source.read_exact(&mut ammo_caliber_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let ammo_caliber_raw = u32::from_be_bytes(ammo_caliber_bytes);
-
-                    let ammo_caliber =
-                        match ammo_caliber_raw {
-                            0 => None,
-                            value => Some(
-                                match object::common::weapons::Caliber::try_from(value) {
-                                    Ok(value) => value,
-                                    Err(_) => return Err(errors::Error::Format(errors::Format::Data))
-                                }
-                            )
-                        };
-
-                    let mut ammo_count_bytes = [0u8; 4];
-                    match source.read_exact(&mut ammo_count_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let ammo_count = u32::from_be_bytes(ammo_count_bytes);
-
-                    let mut ammo_ac_modifier_bytes = [0u8; 4];
-                    match source.read_exact(&mut ammo_ac_modifier_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let ammo_ac_modifier = u32::from_be_bytes(ammo_ac_modifier_bytes);
-
-                    let mut ammo_dr_modifier_bytes = [0u8; 4];
-                    match source.read_exact(&mut ammo_dr_modifier_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let ammo_dr_modifier = u32::from_be_bytes(ammo_dr_modifier_bytes);
-
-                    let mut ammo_dmg_multiplier_bytes = [0u8; 4];
-                    match source.read_exact(&mut ammo_dmg_multiplier_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let ammo_dmg_multiplier = u32::from_be_bytes(ammo_dmg_multiplier_bytes);
-
-                    let mut ammo_dmg_divider_bytes = [0u8; 4];
-                    match source.read_exact(&mut ammo_dmg_divider_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let ammo_dmg_divider = u32::from_be_bytes(ammo_dmg_divider_bytes);
-
-                    object::item::Type::Ammo(
-                        object::item::ammo::Instance {
-                            count: ammo_count,
-                            caliber: ammo_caliber,
-                            adjustments: object::item::ammo::adjustments::Instance {
-                                armor: object::item::ammo::adjustments::Armor {
-                                    class: ammo_ac_modifier,
-                                    resistance: ammo_dr_modifier,
-                                },
-                                damage: object::item::ammo::adjustments::Damage {
-                                    divider: ammo_dmg_divider,
-                                    multiplier: ammo_dmg_multiplier,
-                                },
-                            },
-                        }
-                    )
-                }
-                5 => {
-                    let mut misc_item_pid_bytes = [0u8; 4];
-                    match source.read_exact(&mut misc_item_pid_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let misc_item_pid = u32::from_be_bytes(misc_item_pid_bytes);
-
-                    let mut misc_caliber_bytes = [0u8; 4];
-                    match source.read_exact(&mut misc_caliber_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let misc_caliber_raw = u32::from_be_bytes(misc_caliber_bytes);
-
-                    let misc_caliber =
-                        match misc_caliber_raw {
-                            0 => None,
-                            value => Some(
-                                match object::common::weapons::Caliber::try_from(value) {
-                                    Ok(value) => value,
-                                    Err(_) => return Err(errors::Error::Format(errors::Format::Data))
-                                }
-                            )
-                        };
-
-                    let mut misc_count_bytes = [0u8; 4];
-                    match source.read_exact(&mut misc_count_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let misc_count = u32::from_be_bytes(misc_count_bytes);
-
-                    object::item::Type::Misc(
-                        object::item::misc::Instance {
-                            count: misc_count,
-                            caliber: misc_caliber,
-                            connections: object::item::misc::Connections {
-                                power_item_id: misc_item_pid
-                            },
-                        }
-                    )
-                }
-                6 => {
-                    let mut key_code_bytes = [0u8; 4];
-                    match source.read_exact(&mut key_code_bytes) {
-                        Err(error) => return Err(errors::Error::Read(error)),
-                        Ok(value) => value,
-                    };
-
-                    let key_code = u32::from_be_bytes(key_code_bytes);
-
-                    object::item::Type::Key(
-                        object::item::key::Instance {
-                            code: key_code
-                        }
-                    )
-                }
-                _ => return Err(errors::Error::Format(errors::Format::Type)),
-            };
-
-            object::Type::Item(
-                object::item::Instance {
-                    r#type: item_type,
-                    flags: item_flags,
-                    sprite: match object::common::sprite::Reference::try_from(item_sprite_id_bytes) {
-                        Ok(value) => value,
-                        Err(_) => return Err(errors::Error::Format(errors::Format::Data)),
-                    },
-                    script: match item_script_id_bytes {
-                        [0xFF, 0xFF, 0xFF, 0xFF] => None,
-                        value => match object::common::script::Reference::try_from(value) {
-                            Ok(value) => Some(value),
-                            Err(_) => return Err(errors::Error::Format(errors::Format::Data)),
-                        }
-                    },
-                    actions: item_actions,
-                    material: match object::common::world::Material::try_from(material_id) {
-                        Ok(value) => value,
-                        Err(_) => return Err(errors::Error::Format(errors::Format::Data))
-                    },
-                    size: item_size,
-                    price: item_cost,
-                    weight: item_weight,
-                    connections: object::item::Connections {
-                        _sounds_ids: item_sound_ids,
-                    },
-                }
-            )
-        }
-        // 1 => {}
-        // 2 => {}
-        // 3 => {}
-        // 4 => {}
-        // 5 => {}
-        _ => return Err(errors::Error::Format(errors::Format::Type)),
-    };
-
     Ok(Prototype {
         id: object_id,
-        meta,
-        r#type: object_type,
+        meta: meta::Info {
+            light: meta::info::Light {
+                distance: ScaledValue {
+                    value: lradius,
+                    scale: Range { start: 0, end: 8 },
+                },
+                intensity: ScaledValue {
+                    value: lintensity,
+                    scale: Range { start: 0, end: u16::MAX },
+                },
+            },
+            flags: flagset,
+            sprite: match object::common::sprite::Reference::try_from(sprite_id_bytes) {
+                Ok(value) => value,
+                Err(_) => return Err(errors::Error::Format(errors::Format::Data))
+            },
+            connections: meta::info::Connections { description_id: text_id, },
+        },
+        r#type: match r#type::instance(source, type_id) {
+            Ok(value) => value,
+            Err(error) => return Err(error)
+        }
     })
 }
